Implemented a replay system for viewing arbitrary simulation timepoints.
Simplification of AbilityFactory and ProgramFactory by making them inherit DefaultUidFactory.
Fixed oversight/bug for PlayerData, where it was saved, but never loaded.
Minor addition to Save, to aid in debugging save/load savedata alignment.
Migrated some data from Simulation. Moved the statecopy to the replay manager, and the phase variables to State.
Added Simulation::clear to reset the simulation to a clean state.
Fixed severe bug in World::collide and DataBehaviourT::InternalDatabase::addTypeExt(). Serves me right for not -Walling myself in.
Added replay testing to simtest.cpp and updated it according to the new changes.

+++ simulation/bot/BotCpu.cpp
@@ -73,7 +73,7 @@ namespace Sim {
 		
 		if(newCycles < 0)
 			newCycles = 0;
-		else if(newCycles > mHost->getTypePtr()->cpuStorage)
+		else if(newCycles > int32_t(mHost->getTypePtr()->cpuStorage))
 			newCycles = mHost->getTypePtr()->cpuStorage;
 		
 		mCycleCount = newCycles;
@@ -180,7 +180,7 @@ namespace Sim {
 				if(prog) {
 					// Abort the loop if the CPU can't afford to run the next
 					// program
-					if(mCycleCount >= prog->getCycleCost()) {
+					if(uint32_t(mCycleCount) >= prog->getCycleCost()) {
 						// Note: The CPU immediately eat the cycles
 						// in case the program yields a different cycle
 						// cost after it has processed.
